+++ player.h
@@ -7,9 +7,9 @@ player_t *player_new(uint32_t number, bool golden_possible);
 
 uint32_t player_number(player_t *p);
 
-uint32_t player_busy_fields(player_t *p);
+uint64_t player_busy_fields(player_t *p);
 
-void player_set_busy_fields(player_t *p, uint32_t busy_fields);
+void player_set_busy_fields(player_t *p, uint64_t busy_fields);
 
 uint32_t player_areas(player_t *p);
